Documentation fixes
This fix makes the documented method signatures for `worksheet.write_formula()` and `worksheet.write_array_formula()` match the code, which is left untouched.

It also revises the verbiage and example for including the `value` parameter in `write_array_formula()`. As the functionality is kind of crippled and there is extremely little reason to use the `value` parameter for this method (might as well use `write_formula()` instead), it may be worth removing this parameter from the code, or perhaps just removing it from the docs, at least until some way is devised to write multiple values. (For now, this proposal does not address those decisions.)

Fixes #271.