Do not crash on too long sheet name
The name of a sheet is maximum 31 char long due to an Excel limit.
While the other constrains (invalid character and duplicated) are not possible
to be fixed without user maniupulations, fixing the sheet name size is easy and
does not force the size limitation to be handled by the library user.

Remove the raise Exception and keep only the first 31 characters.

Fixes #364