Eliminate decoration for write methods.
A good portion of execution time was dedicated to
checking wether row and col were integers or
something like A1.  This allows the worksheet to
maintain it's public api while eliminating all of the
needless computation for internal calls.

The increases appear to be nominal, but they
were pretty hefty when running a spreadsheet
with 5m rows through.

      rows       cols    elapsed       size
       144        144       0.66     149608
       144        233       1.12     158128
       144        377       1.72     208816
       233        144       1.05     149632
       233        233       1.68     158128
       233        377       2.76     208864
       377        144       1.79     149680
       377        233       2.86     158104
       377        377       4.57     208840