Fix sizing of objects with position type 1.
Fix the sizing of worksheet objects (images, charts, textboxes)
when the underlying cell sizes have changed and the "object_position"
parameter has been set to 1 "Move and size with cells". An addition
mode 4 has been added to simulate inserting the object before the
cells have changed. Mode 1 will continue to function as if the
object is inserted after the cells have changed.

Issue #618