Allow write_rich_string to accept 2 arguments
I had an issue where `write_rich_string` would not work when the array contained only a format and the string to write in that format:

```
segments = [bold, 'Some bolded string']
worksheet.write_rich_string(row, col, *segments)
```

However, I think this is a legitimate use case and should be allowed. I checked the change in my local repository and it worked, some I am uploading it up here as a pull request.