Write images as `oneCellAnchor` instead of `twoCellAnchor`
Images are currently written as `twoCellAnchor`, which is what Excel uses by default. However, it creates stretching issues when rows are set to auto height. This fixes this issue by writing the images as `oneCellAnchor`, which is supported by the DrawingML specification.

Of course, this is a rather crude integration and some more clever heuristic could be used to determine when to use `oneCellAnchor`, but at least this demonstrates that it can be done and that the resulting files can be opened in Excel.