datetime_to_excel_datetime: adjust the 1900 date system epoch
For most dates, the correct epoch to use is 1899-12-30; only
for dates before 1900-03-01 it is 1899-12-31, due to the leap
year bug. Since most dates encountered will likely be around
the current years, let's go from nearly always having to adjust
to only adjusting if we hit very early dates.

The Office Open XML specification also gives 1899-12-30 as the
base date for the 1900 date system.