Add a type check for the sheetname
If you provide a non-string type as sheetname, e.g. integer or
you end up with an error message, that tells that a variable of
type int does not have a len. Especially for python beginners this
is kind of difficult to understand, so I'd prefer having a clear
message, to what the issue actually is.